novo-projeto-microsservicos/servico_catalogo/catalogo/serializers.py
from rest_framework import serializers
from .models import Categoria, Produto

class CategoriaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Categoria
        fields = ['id', 'nome_categoria']

class ProdutoSerializer(serializers.ModelSerializer):
    categoria = CategoriaSerializer(read_only=True)
    categoria_id = serializers.IntegerField(write_only=True)

    class Meta:
        model = Produto
        fields = ['id', 'nome_produto', 'categoria', 'categoria_id']

