novo-projeto-microsservicos/servico_catalogo/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9

